+++ frontend/src/components/pages/Home.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Home: React.FC = () => {
+    return (
+        <div>
+            <h1 className="text-2xl font-bold">Bienvenido al Sistema de Inventario</h1>
+            <p className="mt-2">Selecciona una opción del menú superior.</p>
+        </div>
+    )
+}
+
+export default Home
